requirements.txt
# ASGO - Adaptive Structured Gradient Optimization
# Installation:
#   For CUDA 11.8: pip install torch>=2.0.0 torchvision --index-url https://download.pytorch.org/whl/cu118
#   For CUDA 12.1: pip install torch>=2.0.0 torchvision --index-url https://download.pytorch.org/whl/cu121
#   Then: pip install -r requirements.txt

# Core ML packages (torch should be installed first with CUDA support, see above)
torch>=2.0.0
torchvision>=0.15.0
transformers==4.56.0
datasets==2.20.0
tiktoken==0.7.0
einops==0.8.0

# Configuration management
hydra-core==1.3.2
omegaconf==2.3.0

# Experiment tracking
wandb==0.21.3

# Scientific computing
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.3.2

# Utilities
tqdm==4.66.4
psutil==6.0.0
pillow==10.4.0
joblib==1.4.2
filelock==3.15.4
PyYAML==6.0.1

# Optional packages (install only if needed)
# torch-geometric  # For graph neural networks
# h5py             # For HDF5 file support